+++ src/zimscraperlib/download.py
@@ -2,9 +2,10 @@
 # -*- coding: utf-8 -*-
 # vim: ai ts=4 sts=4 et sw=4 nu
 
-import subprocess
+import io
 import pathlib
-from concurrent.futures import ThreadPoolExecutor, Future
+import subprocess
+from concurrent.futures import Future, ThreadPoolExecutor
 from typing import Optional, Union
 
 import requests
@@ -120,31 +121,6 @@ class BestMp4(YoutubeConfig):
     }
 
 
-def save_file(
-    url: str,
-    fpath: pathlib.Path,
-    timeout: Optional[int] = 30,
-    retries: Optional[int] = 5,
-) -> requests.structures.CaseInsensitiveDict:
-    """download a file from its URL, and return headers
-
-    Only recommended to be used with small files/HTMLs"""
-
-    for left_attempts in range(retries, -1, -1):
-        try:
-            resp = requests.get(url, timeout=timeout)
-            resp.raise_for_status()
-            with open(fpath, "wb") as fp:
-                fp.write(resp.content)
-            return resp.headers
-        except requests.exceptions.RequestException as exc:
-            logger.debug(
-                f"Request for {url} failed ({left_attempts} attempts left)\n{exc}"
-            )
-            if left_attempts == 0:
-                raise exc
-
-
 def save_large_file(url: str, fpath: pathlib.Path) -> None:
     """ download a binary file from its URL, using wget """
     subprocess.run(
@@ -162,3 +138,78 @@ def save_large_file(url: str, fpath: pathlib.Path) -> None:
         ],
         check=True,
     )
+
+
+def stream_file(
+    url: str,
+    fpath: Optional[pathlib.Path] = None,
+    byte_stream: Optional[io.BytesIO] = None,
+    block_size: Optional[int] = 1024,
+    proxies: Optional[dict] = None,
+    only_first_block: Optional[bool] = False,
+    max_retries: Optional[int] = 5,
+) -> Union[int, requests.structures.CaseInsensitiveDict]:
+    """Stream data from a URL to either a BytesIO object or a file
+    Arguments -
+        fpath - Path of the file where data is sent
+        byte_stream - The BytesIO object where data is sent
+        block_size - Size of each chunk of data read in one iteration
+        proxies - A dict of proxies to be used (More here - https://requests.readthedocs.io/en/master/user/advanced/#proxies)
+        only_first_block - Whether to download only one (first) block
+        max_retries - Maximum number of retries after which error is raised
+    Returns the total number of bytes downloaded and the response headers"""
+
+    # if no output option is supplied
+    if fpath is None and byte_stream is None:
+        raise ValueError("Either file path or a bytesIO object is needed")
+
+    # prepare adapter so it retries on failure
+    session = requests.Session()
+    # retries up-to FAILURE_RETRIES whichever kind of listed error
+    retries = requests.packages.urllib3.util.retry.Retry(
+        total=max_retries,  # total number of retries
+        connect=max_retries,  # connection errors
+        read=max_retries,  # read errors
+        status=2,  # failure HTTP status (only those bellow)
+        redirect=False,  # don't fail on redirections
+        backoff_factor=30,  # sleep factor between retries
+        status_forcelist=[
+            413,
+            429,
+            500,
+            502,
+            503,
+            504,
+        ],  # force retry on the following codes
+    )
+
+    retry_adapter = requests.adapters.HTTPAdapter(max_retries=retries)
+    session.mount("http", retry_adapter)  # tied to http and https
+    resp = session.get(
+        url,
+        stream=True,
+        proxies=proxies,
+    )
+    resp.raise_for_status()
+
+    total_downloaded = 0
+    if fpath is not None:
+        fp = open(fpath, "wb")
+    else:
+        fp = byte_stream
+
+    for data in resp.iter_content(block_size):
+        total_downloaded += len(data)
+        fp.write(data)
+
+        # stop downloading/reading if we're just testing first block
+        if only_first_block:
+            break
+
+    logger.info(f"Downloaded {total_downloaded} bytes from {url}")
+
+    if fpath:
+        fp.close()
+    else:
+        fp.seek(0)
+    return total_downloaded, resp.headers