+++ tests/download/test_download.py
@@ -2,13 +2,14 @@
 # -*- coding: utf-8 -*-
 # vim: ai ts=4 sts=4 et sw=4 nu
 
+import io
 import pytest
 import requests
 import pathlib
 import concurrent.futures
 
 from zimscraperlib.download import (
-    save_file,
+    stream_file,
     save_large_file,
     YoutubeDownloader,
     BestWebm,
@@ -34,20 +35,36 @@ def get_dest_file(tmp_path):
 
 
 def test_missing_dest(tmp_path):
-    with pytest.raises(TypeError):
-        save_file("http://some_url")
+    with pytest.raises(requests.exceptions.ConnectionError):
+        stream_file(url="http://some_url", byte_stream=io.BytesIO())
 
 
 def test_invalid_url(tmp_path, invalid_url):
     dest_file = tmp_path / "favicon.ico"
     with pytest.raises(requests.exceptions.ConnectionError):
-        save_file(invalid_url, dest_file)
+        stream_file(url=invalid_url, fpath=dest_file)
+
+
+def test_no_output_supplied(valid_http_url):
+    with pytest.raises(
+        ValueError, match="Either file path or a bytesIO object is needed"
+    ):
+        stream_file(url=valid_http_url)
+
+
+def test_first_block_download(valid_http_url):
+    byte_stream = io.BytesIO()
+    size, ret = stream_file(
+        url=valid_http_url, byte_stream=byte_stream, only_first_block=True
+    )
+    assert_headers(ret)
+    assert len(byte_stream.read()) == 3062
 
 
 @pytest.mark.slow
 def test_save_http(tmp_path, valid_http_url):
     dest_file = tmp_path / "favicon.ico"
-    ret = save_file(valid_http_url, dest_file)
+    size, ret = stream_file(url=valid_http_url, fpath=dest_file)
     assert_headers(ret)
     assert_downloaded_file(valid_http_url, dest_file)
 
@@ -55,30 +72,31 @@ def test_save_http(tmp_path, valid_http_url):
 @pytest.mark.slow
 def test_save_https(tmp_path, valid_https_url):
     dest_file = tmp_path / "favicon.ico"
-    ret = save_file(valid_https_url, dest_file)
+    size, ret = stream_file(url=valid_https_url, fpath=dest_file)
     assert_headers(ret)
     assert_downloaded_file(valid_https_url, dest_file)
+
+
+@pytest.mark.slow
+def test_stream_to_bytes(valid_https_url):
+    byte_stream = io.BytesIO()
+    size, ret = stream_file(url=valid_https_url, byte_stream=byte_stream)
+    assert_headers(ret)
+    assert byte_stream.read() == requests.get(valid_https_url).content
 
 
 @pytest.mark.slow
 def test_save_parent_folder_missing(tmp_path, valid_http_url):
     dest_file = tmp_path / "some-folder" / "favicon.ico"
     with pytest.raises(IOError):
-        save_file(valid_http_url, dest_file)
+        stream_file(url=valid_http_url, fpath=dest_file)
 
 
 @pytest.mark.slow
 def test_save_http_error(tmp_path, http_error_url):
     dest_file = tmp_path / "favicon.ico"
     with pytest.raises(requests.exceptions.HTTPError):
-        save_file(http_error_url, dest_file)
-
-
-@pytest.mark.slow
-def test_save_timeout(tmp_path, timeout_url):
-    dest_file = tmp_path / "favicon.ico"
-    with pytest.raises(requests.exceptions.RequestException):
-        save_file(timeout_url, dest_file, timeout=1)
+        stream_file(url=http_error_url, fpath=dest_file)
 
 
 @pytest.mark.slow