+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.3.5.dev0
 
-* n/a
+* add `stream_file()` to stream content from a URL into a file or a `BytesIO` object
+* deprecated `save_file()`
 
 # 1.3.4